Rename Streamly.Internal.Data.Unicode.Stream
According to the new module structure, it should be Streamly.Internal.Unicode.Stream
