+++ test/String.hs
@@ -13,7 +13,7 @@ import qualified Streamly.Memory.Array as A
 import qualified Streamly.Internal.Memory.ArrayStream as AS
 import qualified Streamly.Prelude as S
 import qualified Streamly.Data.Unicode.Stream as SS
-import qualified Streamly.Internal.Data.Unicode.Stream as IUS
+import qualified Streamly.Internal.Unicode.Stream as IUS
 import qualified Streamly.Internal.Memory.Unicode.Array as IUA
 
 -- Coverage build takes too long with default number of tests
